relocate arrayToGrid
