+++ src/logic/convertRepresentativeStringToGrid.js
@@ -1,4 +1,4 @@
-import {arrayToGrid} from "./arrayToGrid";
+import {arrayToGrid} from "../common/arrayToGrid";
 import {cipherLetter} from "./cipherLetter";
 
 // Converts a string of letters and integers into a square grid of single letters and empty strings.
